feat: admin에 에러 추가

See merge request sw_track/class_05/web_project/team03/back-end!69

+++ src/services/adminService.js
@@ -1,17 +1,24 @@
 const { productModel } = require('../db/models');
 const { orderModel } = require('../db/models');
 const { categoryModel } = require('../db/models');
+const { userModel } = require('../db/models')
 const AppError = require('../misc/AppError');
 
 class AdminService {
-  constructor(productModel, orderModel, categoryModel) {
+  constructor(productModel, orderModel, categoryModel, userModel) {
     this.productModel = productModel;
     this.orderModel = orderModel;
     this.categoryModel = categoryModel;
+    this.userModel = userModel;
   }
 
   // 상품 등록 관리자
   async postProduct(product) {
+    const { id } = product;
+    const Product = await this.productModel.findOne(id);
+    if (Product) {
+      throw new AppError('Bad Request', 400, '이미 존재하는 상품입니다.');
+    }
     const result = await this.productModel.create(product);
     return result;
   }
@@ -19,17 +26,30 @@ class AdminService {
   // 상품 단건 조회 관리자
   async getProduct(productId) {
     const product = await this.productModel.findOne(productId);
+    if (!product) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 상품입니다.');
+    }
     return product;
   }
 
   // 상품 단건 수정 관리자
   async putProduct(product) {
+    const { id } = product;
+    const Product = await this.productModel.findOne(id);
+    if (!Product) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 상품입니다.');
+    }
     const result = await this.productModel.update(product);
     return result;
   }
 
   // 상품 단건 삭제 관리자
   async deleteProduct(productId) {
+    const { id } = product;
+    const Product = await this.productModel.findOne(id);
+    if (!Product) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 상품입니다.');
+    }
     const result = await this.productModel.deleteOne(productId);
     return result;
   }
@@ -42,12 +62,20 @@ class AdminService {
 
   // 주문 내역 특정 사용자 조회 관리자
   async getOrder(id) {
+    const user = await this.userModel.findbyId(id);
+    if (!user) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 사용자입니다.');
+    }
     const order = await this.orderModel.findOrdersByUserId(id);
     return order;
   }
 
   // 주문 내역 삭제 관리자
   async deleteOrder(id) {
+    const user = await this.userModel.findbyId(id);
+    if (!user) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 사용자입니다.');
+    }
     const result = await this.orderModel.deleteOrder(id);
     return result;
   }
@@ -97,6 +125,10 @@ class AdminService {
 
   // 주문 내역 주문상태 변경 관리자
   async putStatus(id, status) {
+    const order = await this.orderModel.findOrderByOrderId(id);
+    if (!order) {
+      throw new AppError('Bad Request', 400, '존재하지 않는 주문입니다.')
+    }
     const result = await this.orderModel.updateStatus(id, status);
     return result;
   }
